refactor: split Constantine signed-Booth recoder into pippenger_constantine.hpp
Move the ~360 lines of Constantine slice params, scalar packed-digit recoder,
and SIMD x4 specialisations out of scalar_multiplication.cpp into a dedicated
pippenger_constantine.hpp under the same round_parallel_detail namespace.
Pure code motion.
